elfloader: arm: fix function declaration (type mismatch)
This chops off the aff3 level on Aarch64.

Why does the compiler not warn??? Because the own header was not
included. If you just include the header (without the change of
the return value in the header), we get:

/home/mro/nvos_neu2/tools/seL4_tools/elfloader-tool/src/arch-arm/64/cpuid.c:14:10: error: conflicting types for 'read_cpuid_mpidr'
   14 | uint32_t read_cpuid_mpidr(void)
      |          ^~~~~~~~~~~~~~~~
In file included from /home/mro/nvos_neu2/tools/seL4_tools/elfloader-tool/src/arch-arm/64/cpuid.c:9:
/home/mro/nvos_neu2/tools/seL4_tools/elfloader-tool/include/arch-arm/cpuid.h:15:8: note: previous declaration of 'read_cpuid_mpidr' was here
   15 | word_t read_cpuid_mpidr(void);
      |        ^~~~~~~~~~~~~~~~
[190/200] Building C object elfloader-tool/CMakeFiles/elfloader.dir/src/arch-arm/smp_boot.c.obj
ninja: build stopped: subcommand failed.

Signed-off-by: Matthias Rosenfelder <matthias.rosenfelder@nio.io>