elfloader: setup pagetables as needed
This change sets up pagetables individually for:
- The ELFloader image (Normal memory)
- The DTB, whether supplied by EFI, cpio or u-boot (Normal mem)
- The UART MMIO range (Strongly-Ordered mem)

Thus, it removes the bulk 512 GiB 1:1 mapping that was there before.
This resulted in problems, since the kernel image was mapped with
Normal memory, but the same physical memory was part of the
1:1 Strongly-Ordered mapping.

This fulfills the definition of "Mismatched memory attributes"
from the ARM Architecture specification (ARM DDI 0487I.a, section
B2.8). Even though I am currently unable to see where there
would *occur* such a mismatched access, having such a mapping
situation is certainly not desirable and should be avoided.

Moreover, it is unclear whether there could arise problems from
establishing the (Strongly-ordered) mapping if there is nothing
behind a physical address (which is certainly true for some
parts of the 512 GiB range).

This commit solves the sporadics hangs while booting after the
"Enabling MMU and ..." message. Tests on several different Orins
(Muc and SJ) show promising results, i.e. no "hangs" occurred
anymore.

Note: The code in arm_switch_to_hyp_tables() still disables and
re-enables both MMU & caches, but there are no memory accesses in
between. That section has been engineered to be as short as possible
and no memory accesses happen in between.

Several barriers and code to invalidate instruction caches have
been added, too, in order to be on the safe side. However, tests
with just adding *that* code still showed the problem being present.
The only change that showed behavior change was the change of
translation tables. Thus, this *is* the actual solution to the
instability problems.

Moreover, we need to support crossing a 1 GiB page for placement
of the ELFloader. This is due to the latest firmware on Orin0 in
MUC, named "Jetson UEFI firmware (version 4.1-33958178)", which puts
our image closely below a 1 GiB boundary. Only for tiny image sizes
the boundary will not be crossed. Thus, we do not hard-code the
writing of tables, because the logic for doing so while crossing a
1 GiB boundary is too complicated. Instead, we use a fully dynamic
approach that walks the pagetables in software for a given VA and
inserts missing levels on demand from a preallocated pool of pages.

Only the two top-level pagetables are fixed. This allows for re-use
of all pagetable code, where we only need to distinguish in one (!)
place between hypervisor and non-hyp (or VHE).

Signed-off-by: Matthias Rosenfelder <matthias.rosenfelder@nio.io>