elfloader: arm: flush dcache by va range instead of set/way
Cleaning a cache by set/way only cleans the CPU-local caches. System
caches can only be cleaned by virtual address, thus the elfloader
is changed to clean the data cache by VA instead of set/way.

This is important as the point-of-coherency (PoC) may lie beyond the
system cache, in which case we cannot clean to PoC unless we use a clean
by va.

Co-authored-by: Matthias Rosenfelder <matthias.rosenfelder@nio.io>
Co-authored-by: Borna Blazevic <borna.blazevic@nio.io>
Signed-off-by: Andy Bui <andy.bui@nio.io>